[Rename] `storage.ts` to `diskStorage.ts`
