+++ src/code/platform/configuration/common/configRegistrant.ts
@@ -1,0 +1,28 @@
+
+/**
+ * An interface only for {@link ConfigRegistrant}.
+ */
+export interface IConfigurationRegistrant {
+
+}
+
+/**
+ * // TODO
+ */
+export const ConfigRegistrant = new class implements IConfigurationRegistrant {
+
+    // [field]
+
+
+
+    // [constructor]
+
+    constructor() {
+
+    }
+
+    // [public methods]
+
+
+
+}
