+++ src/code/platform/files/common/diskStorage.ts
