+++ src/code/platform/status/electron/mainStatusService.ts
@@ -7,7 +7,7 @@ import { NOTA_DIR_NAME } from "src/code/platform/configuration/electron/configSe
 import { IFileService } from "src/code/platform/files/common/fileService";
 import { createDecorator } from "src/code/platform/instantiation/common/decorator";
 import { IEnvironmentService, IMainEnvironmentService } from "src/code/platform/environment/common/environment";
-import { DiskStorage, IDiskStorage } from "src/code/platform/files/common/storage";
+import { DiskStorage, IDiskStorage } from "src/code/platform/files/common/diskStorage";
 import { IMainLifeCycleService } from "src/code/platform/lifeCycle/electron/mainLifeCycleService";
 
 export const IMainStatusService = createDecorator<IMainStatusService>('status-service');
