+++ test/code/platform/configStorage.test.ts
